fix sqlApi script
