+++ ppml/trusted-bigdata/examples/sql_functions_f_example.py
@@ -1,66 +1,93 @@
 from pyspark.sql import SparkSession
 from pyspark.sql.functions import *
-from pyspark.sql.types import *
 
-def sql_functions_f_example(spark):
-
-    # factorial
-    df = spark.createDataFrame([(5,)], ['n'])
-    df.select(factorial(df.n).alias('f')).show()
-    print("factorial API finished")
-
-    # first
-    df = spark.createDataFrame([(None,), (5,), (6,)], ['n'])
-    df.select(first(df.n)).show()
-    print("first API finished")
-
-    # flatten
-    #the result is null if there has 'None' in arrays
-    df = spark.createDataFrame([([[1, 2, 3], [4, 5], [6]],), ([None, [4, 5]],)], ['data'])
-    df.select(flatten(df.data).alias('r')).show()
-    print("flatten API finished")
-
-    # floor
-    df = spark.createDataFrame([(1.2,), (5.9,), (6.3,)], ['n'])
-    df.select(floor(df.n)).show()
-    print("floor API finished")
-
-    # format_number
-    spark.createDataFrame([(5,)], ['a']).select(format_number('a', 4).alias('v')).show()
-    print("format_number API finished")
+def sql_functions_d_example(spark):
     
-    # format_string
-    df = spark.createDataFrame([(5, "hello")], ['a', 'b'])
-    df.select(format_string('%d %s', df.a, df.b).alias('v')).show()
-    print("format_string API finished")
+    # date_add
+    df = spark.createDataFrame([('2015-04-08',)], ['dt'])
+    df.select(date_add(df.dt,1).alias("next_day")).show()
+    print("date_add API finished")
 
-    # from_json
-    data = [(1, '''{"a": 1}''')]
-    schema = StructType([StructField("a", IntegerType())])
-    df = spark.createDataFrame(data, ("key", "value"))
-    df.select(from_json(df.value, schema).alias("json")).show()
-    df.select(from_json(df.value, "MAP<STRING,INT>").alias("json")).show()
-    print("from_json API finished")
+    # date_format
+    df = spark.createDataFrame([('2015-04-08',)], ['dt'])
+    df.select(date_format('dt', 'MM/dd/yyy').alias('date')).show()
+    print("date_format API finished")
 
-    # from_unixtime
-    time_df = spark.createDataFrame([(1428476400,)], ['unix_time'])
-    time_df.select(from_unixtime('unix_time').alias('ts')).show()
-    print("from_unixtime API finished")
+    # date_sub
+    df = spark.createDataFrame([('2015-04-08',)], ['dt'])
+    df.select(date_sub(df.dt, 1).alias('prev_date')).show()
+    print("date_sub API finished")
 
-    # from_utc_timestamp
-    df = spark.createDataFrame([('1997-02-28 10:30:00', 'JST')], ['ts', 'tz'])
-    df.select(from_utc_timestamp(df.ts, "PST").alias('local_time')).show()
-    df.select(from_utc_timestamp(df.ts, df.tz).alias('local_time')).show()
-    print("from_utc_timestamp API finished")
+    # date_trunc
+    df = spark.createDataFrame([('1997-02-28 05:02:11',)], ['t'])
+    df.select(date_trunc('year', df.t).alias('year')).show()
+    df.select(date_trunc('mon', df.t).alias('month')).show()
+    print("date_trunc API finished")
+
+    # datediff
+    #arg1:end time / arg2:start time
+    #Returns the number of days from start to end
+    df = spark.createDataFrame([('2015-04-08','2015-05-10')], ['d1', 'd2'])
+    df.select(datediff(df.d2, df.d1).alias('diff')).show()
+    print("datediff API finished")
+
+    # dayofmonth
+    df = spark.createDataFrame([('2015-04-08',)], ['dt'])
+    df.select(dayofmonth('dt').alias('day')).show()
+    print("dayofmonth API finished")
+
+    # dayofweek
+    df = spark.createDataFrame([('2015-04-08',)], ['dt'])
+    df.select(dayofweek('dt').alias('day')).show()
+    print("dayofweek API finished")
+
+    # dayofyear
+    df = spark.createDataFrame([('2015-04-08',)], ['dt'])
+    df.select(dayofyear('dt').alias('day')).show()
+    print("dayofyear API finished")
+
+    # decode
+    df = spark.createDataFrame([(1, 'a'), (2, 'b'), (3, 'c')], ["n1", "s1"])
+    df.withColumn("encode", encode(df.s1, "utf-8")).withColumn("decode", decode("encode", "utf-8")).show()
+    print("decode API finished")
+
+    # degrees
+    import math
+    df = spark.createDataFrame([(math.pi,), (math.pi / 6,)], ["radians"])
+    df.select(degrees(df.radians)).show()
+    print("degrees API finished")
+
+    # dense_rank
+    from pyspark.sql import Window
+    window = Window.orderBy("score")
+    df = spark.createDataFrame([('Bob', 90), ('Alice', 95), ('Coris', 90), ('David', 89)], ["name", "score"])
+    df.withColumn("dense_rank", dense_rank().over(window)).show()
+    df.withColumn("rank", rank().over(window)).show()
+    print("dense_rank API finished")
+
+    # desc
+    df = spark.createDataFrame([(1, None), (10, 12), (8, 3), (None, 9), (9, 6)], ["n1", "n2"])
+    df.sort(df.n1.desc()).show()
+    print("desc API finished")
+
+    # desc_nulls_first
+    df = spark.createDataFrame([(1, None), (10, 12), (8, 3), (None, 9), (9, 6)], ["n1", "n2"])
+    df.sort(df.n1.desc_nulls_first()).show()
+    print("desc_nulls_first API finished")
+
+    # desc_nulls_last
+    df = spark.createDataFrame([(1, None), (10, 12), (8, 3), (None, 9), (9, 6)], ["n1", "n2"])
+    df.sort(df.n1.desc_nulls_last()).show()
+    print("desc_nulls_last API finished")
     
-    print("Finish running function_f API")
+    print("Finish running function_d API")
 
 if __name__ == "__main__":
 
     spark = SparkSession \
         .builder \
-        .appName("Python Spark SQL functions_f API example") \
+        .appName("Python Spark SQL functions_d API example") \
         .config("spark.some.config.option", "some-value") \
         .getOrCreate()
-    sql_functions_f_example(spark)
+    sql_functions_d_example(spark)
     spark.stop()
